Moved AccessRestrictions to `/discord`
