+++ ts_src/bot/modules/index.ts
@@ -1,5 +1,4 @@
 
 export * from './ping.module';
-export * from './speedrun.module';
+export * from './ylsrcom.module';
 export * from './links.module';
-export * from './restricted.module';
