+++ ts_src/discord/index.ts
@@ -2,3 +2,4 @@ export * from './arguments';
 export * from './user';
 export * from './channel';
 export * from './permissions';
+export * from './accessrestrictions';
